Fix overcomplicated journal dir handling
It's much simpler to just expand the classic tilde, no need for special character bs or forced saves upon first boot. Just fill the profile path when settings are saved, and fill it if the tilde (~) is ever passed to the Python program. I really overcomplicated that the first time around.